feat: Rename social.core to social package - Meeds-io/MIPs#165
