+++ component/api/src/main/java/io/meeds/social/richeditor/RichEditorConfiguration.java
@@ -1,25 +1,31 @@
 /*
  * This file is part of the Meeds project (https://meeds.io/).
- *
- * Copyright (C) 2020 - 2024 Meeds Association contact@meeds.io
- *
+ * Copyright (C) 2020 - 2023 Meeds Association contact@meeds.io
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 3 of the License, or (at your option) any later version.
- *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
  * Lesser General Public License for more details.
- *
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- *
  */
-package io.meeds.social.core.space.constant;
+package io.meeds.social.richeditor;
 
-public enum Registration {
-  OPEN, VALIDATION, CLOSED
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+public class RichEditorConfiguration {
+
+  private String filePath;
+
+  private String instanceType;
+
 }
