+++ component/api/src/main/java/io/meeds/social/richeditor/RichEditorConfigurationPlugin.java
@@ -13,7 +13,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package io.meeds.social.core.richeditor;
+package io.meeds.social.richeditor;
 
 import java.util.ArrayList;
 import java.util.Iterator;
