+++ component/api/src/main/java/io/meeds/social/richeditor/RichEditorConfigurationService.java
@@ -1,25 +1,39 @@
 /*
  * This file is part of the Meeds project (https://meeds.io/).
- *
- * Copyright (C) 2020 - 2024 Meeds Association contact@meeds.io
- *
+ * 
+ * Copyright (C) 2020 - 2023 Meeds Association contact@meeds.io
+ * 
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 3 of the License, or (at your option) any later version.
- *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
  * Lesser General Public License for more details.
- *
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- *
  */
-package io.meeds.social.core.space.constant;
+package io.meeds.social.richeditor;
 
-public enum Visibility {
-  PUBLIC, PRIVATE, HIDDEN
+public interface RichEditorConfigurationService {
+
+  /**
+   * Retrieves the rich editor configuration file for a given instance type
+   * 
+   * @param  instanceType Rich editor instance type
+   * @return              Javascript content of rich Editor configuration
+   */
+  String getRichEditorConfiguration(String instanceType);
+
+  /**
+   * Add a new plugin defining a part of Rich editor instance configuration.
+   * If the plugin contains an Object Parameter that doesn't define instanceType,
+   * the configuration will go to all Rich Editor instances.
+   * 
+   * @param plugin {@link RichEditorConfigurationPlugin}
+   */
+  void addPlugin(RichEditorConfigurationPlugin plugin);
+
 }
