+++ component/core/src/main/java/io/meeds/social/search/listener/SpaceIndexingListenerImpl.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package io.meeds.social.core.search.listener;
+package io.meeds.social.search.listener;
 
 import java.util.Set;
