+++ component/core/src/main/java/io/meeds/social/search/model/SpaceSearchResult.java
@@ -7,41 +7,32 @@
  * modify it under the terms of the GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 3 of the License, or (at your option) any later version.
- *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
  * Lesser General Public License for more details.
  *
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package io.meeds.social.core.search.model;
+package io.meeds.social.search.model;
 
 import java.util.List;
 
-import io.meeds.social.space.constant.SpaceMembershipStatus;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 
+@NoArgsConstructor
 @AllArgsConstructor
-@EqualsAndHashCode
 @Data
-public class SpaceSearchFilter {
+public class SpaceSearchResult {
 
-  private String                username;
+  private long         id;
 
-  private long                  userIdentityId;
+  private List<String> nameExcerpts;
 
-  private String                term;
-
-  private boolean               favorites;
-
-  private List<String>          tagNames;
-
-  private SpaceMembershipStatus status;
+  private List<String> descriptionExcerpts;
 
 }
