+++ component/core/src/main/java/io/meeds/social/space/constant/Registration.java
@@ -18,8 +18,8 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
  */
-package io.meeds.social.core.space.constant;
+package io.meeds.social.space.constant;
 
-public enum PublicSiteVisibility {
-  MANAGER, MEMBER, INTERNAL, AUTHENTICATED, EVERYONE;
+public enum Registration {
+  OPEN, VALIDATION, CLOSED
 }
