+++ component/core/src/main/java/io/meeds/social/space/listener/SpaceLayoutHandlerPlugin.java
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package io.meeds.social.core.space.listener;
+package io.meeds.social.space.listener;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -27,7 +27,7 @@ import org.exoplatform.social.core.space.spi.SpaceLifeCycleEvent;
 import org.exoplatform.social.core.space.spi.SpaceLifeCycleListener;
 import org.exoplatform.social.core.space.spi.SpaceService;
 
-import io.meeds.social.core.space.service.SpaceLayoutService;
+import io.meeds.social.space.service.SpaceLayoutService;
 
 import jakarta.annotation.PostConstruct;
