+++ component/core/src/main/java/io/meeds/social/space/service/SpaceLayoutService.java
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package io.meeds.social.core.space.service;
+package io.meeds.social.space.service;
 
 import java.util.List;
