+++ component/core/src/main/java/org/exoplatform/social/core/jpa/storage/entity/SpaceEntity.java
@@ -24,9 +24,9 @@ import java.util.Set;
 
 import org.exoplatform.commons.utils.StringListConverter;
 
-import io.meeds.social.core.space.constant.PublicSiteVisibility;
-import io.meeds.social.core.space.constant.Registration;
-import io.meeds.social.core.space.constant.Visibility;
+import io.meeds.social.space.constant.PublicSiteVisibility;
+import io.meeds.social.space.constant.Registration;
+import io.meeds.social.space.constant.Visibility;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
