+++ component/core/src/test/java/io/meeds/social/space/service/SpaceLayoutServiceTest.java
@@ -33,7 +33,7 @@ import org.exoplatform.social.core.space.SpaceUtils;
 import org.exoplatform.social.core.space.model.Space;
 import org.exoplatform.social.core.test.AbstractCoreTest;
 
-import io.meeds.social.core.space.service.SpaceLayoutService;
+import io.meeds.social.space.service.SpaceLayoutService;
 
 import lombok.SneakyThrows;
