+++ component/core/src/test/java/io/meeds/social/space/template/storage/SpaceTemplateStorageMock.java
@@ -24,8 +24,8 @@ import java.util.List;
 
 import org.springframework.data.domain.Pageable;
 
-import io.meeds.social.core.space.constant.Registration;
-import io.meeds.social.core.space.constant.Visibility;
+import io.meeds.social.space.constant.Registration;
+import io.meeds.social.space.constant.Visibility;
 import io.meeds.social.space.template.model.SpaceTemplate;
 
 public class SpaceTemplateStorageMock extends SpaceTemplateStorage {
