+++ component/core/src/test/java/org/exoplatform/social/core/jpa/storage/dao/SpaceDAOTest.java
@@ -31,9 +31,9 @@ import org.exoplatform.social.core.jpa.storage.entity.SpaceEntity;
 import org.exoplatform.social.core.jpa.storage.entity.SpaceMemberEntity;
 import org.exoplatform.social.core.jpa.test.BaseCoreTest;
 
-import io.meeds.social.core.space.constant.Registration;
-import io.meeds.social.core.space.constant.Visibility;
+import io.meeds.social.space.constant.Registration;
 import io.meeds.social.space.constant.SpaceMembershipStatus;
+import io.meeds.social.space.constant.Visibility;
 
 public class SpaceDAOTest extends BaseCoreTest {
   private SpaceDAO    spaceDAO;
