+++ component/core/src/test/java/org/exoplatform/social/core/test/NoContainerTestSuite.java
@@ -35,8 +35,8 @@ import org.exoplatform.social.core.service.GettingStartedServiceTest;
 import org.exoplatform.social.core.storage.StorageUtilsTest;
 
 import io.meeds.social.authorization.AuthorizationManagerTest;
-import io.meeds.social.core.search.SpaceSearchConnectorTest;
 import io.meeds.social.image.plugin.ImageAttachmentPluginTest;
+import io.meeds.social.search.SpaceSearchConnectorTest;
 import io.meeds.social.space.template.plugin.attachment.SpaceTemplateBannerAttachmentPluginTest;
 import io.meeds.social.space.template.plugin.translation.SpaceTemplateTranslationPluginTest;
 import io.meeds.social.space.template.service.SpaceTemplateServiceTest;
