+++ component/core/src/test/resources/conf/exo.social.component.core-configuration.xml
@@ -41,7 +41,7 @@
   </component>
 
   <component>
-    <type>io.meeds.social.core.search.SpaceSearchConnector</type>
+    <type>io.meeds.social.search.SpaceSearchConnector</type>
     <init-params>
       <value-param>
         <name>query.file.path</name>
@@ -342,11 +342,11 @@
 
   <component>
     <key>org.exoplatform.social.core.space.spi.SpaceService</key>
-    <type>io.meeds.social.core.space.service.SpaceServiceImpl</type>
+    <type>io.meeds.social.space.service.SpaceServiceImpl</type>
   </component>
 
   <component>
-    <type>io.meeds.social.core.space.service.SpaceLayoutService</type>
+    <type>io.meeds.social.space.service.SpaceLayoutService</type>
   </component>
 
   <component>
@@ -1472,7 +1472,7 @@
     <component-plugin>
       <name>SpacePublicSiteListener</name>
       <set-method>addSpaceListener</set-method>
-      <type>io.meeds.social.core.space.listener.SpacePublicSiteListener</type>
+      <type>io.meeds.social.space.listener.SpacePublicSiteListener</type>
     </component-plugin>
   </external-component-plugins>
 
@@ -1481,7 +1481,7 @@
     <component-plugin>
       <name>SpaceIndexingListenerImpl</name>
       <set-method>addSpaceListener</set-method>
-      <type>io.meeds.social.core.search.listener.SpaceIndexingListenerImpl</type>
+      <type>io.meeds.social.search.listener.SpaceIndexingListenerImpl</type>
     </component-plugin>
   </external-component-plugins>
