+++ component/service/src/main/java/org/exoplatform/social/rest/impl/space/SpaceRest.java
@@ -95,7 +95,7 @@ import org.exoplatform.web.login.recovery.PasswordRecoveryService;
 
 import io.meeds.portal.security.constant.UserRegistrationType;
 import io.meeds.portal.security.service.SecuritySettingService;
-import io.meeds.social.core.space.service.SpaceLayoutService;
+import io.meeds.social.space.service.SpaceLayoutService;
 import io.meeds.social.util.JsonUtils;
 
 import io.swagger.v3.oas.annotations.Operation;
