+++ webapp/src/main/webapp/WEB-INF/conf/social-extension/social/ckeditor-configuration.xml
@@ -24,7 +24,7 @@
    xmlns="http://www.exoplatform.org/xml/ns/kernel_1_3.xsd">
 
   <component>
-    <key>io.meeds.social.core.richeditor.RichEditorConfigurationService</key>
+    <key>io.meeds.social.richeditor.RichEditorConfigurationService</key>
     <type>io.meeds.social.richeditor.RichEditorConfigurationServiceImpl</type>
   </component>
 
@@ -33,15 +33,15 @@
   </component>
 
   <external-component-plugins>
-    <target-component>io.meeds.social.core.richeditor.RichEditorConfigurationService</target-component>
+    <target-component>io.meeds.social.richeditor.RichEditorConfigurationService</target-component>
     <component-plugin>
       <name>BaseCKEditorConfiguration</name>
       <set-method>addPlugin</set-method>
-      <type>io.meeds.social.core.richeditor.RichEditorConfigurationPlugin</type>
+      <type>io.meeds.social.richeditor.RichEditorConfigurationPlugin</type>
       <init-params>
         <object-param>
           <name>BaseCKEditorConfiguration</name>
-          <object type="io.meeds.social.core.richeditor.RichEditorConfiguration">
+          <object type="io.meeds.social.richeditor.RichEditorConfiguration">
             <field name="instanceType">
               <string></string>
             </field>
@@ -55,11 +55,11 @@
     <component-plugin>
       <name>BaseCKEditorConfiguration</name>
       <set-method>addPlugin</set-method>
-      <type>io.meeds.social.core.richeditor.RichEditorConfigurationPlugin</type>
+      <type>io.meeds.social.richeditor.RichEditorConfigurationPlugin</type>
       <init-params>
         <object-param>
           <name>BaseCKEditorConfiguration</name>
-          <object type="io.meeds.social.core.richeditor.RichEditorConfiguration">
+          <object type="io.meeds.social.richeditor.RichEditorConfiguration">
             <field name="instanceType">
               <string>activityContent</string>
             </field>
@@ -73,11 +73,11 @@
     <component-plugin>
       <name>BaseCKEditorConfiguration</name>
       <set-method>addPlugin</set-method>
-      <type>io.meeds.social.core.richeditor.RichEditorConfigurationPlugin</type>
+      <type>io.meeds.social.richeditor.RichEditorConfigurationPlugin</type>
       <init-params>
         <object-param>
           <name>BaseCKEditorConfiguration</name>
-          <object type="io.meeds.social.core.richeditor.RichEditorConfiguration">
+          <object type="io.meeds.social.richeditor.RichEditorConfiguration">
             <field name="instanceType">
               <string>activityComment</string>
             </field>
