+++ webapp/src/main/webapp/WEB-INF/conf/social-extension/social/core-configuration.xml
@@ -397,7 +397,7 @@
 
     <component>
       <key>org.exoplatform.social.core.space.spi.SpaceService</key>
-      <type>io.meeds.social.core.space.service.SpaceServiceImpl</type>
+      <type>io.meeds.social.space.service.SpaceServiceImpl</type>
     </component>
 
     <component>
@@ -723,7 +723,7 @@
       <component-plugin>
         <name>SpacePublicSiteListener</name>
         <set-method>addSpaceListener</set-method>
-        <type>io.meeds.social.core.space.listener.SpacePublicSiteListener</type>
+        <type>io.meeds.social.space.listener.SpacePublicSiteListener</type>
       </component-plugin>
     </external-component-plugins>
